requirements.txt
Flask==2.3.3
openai==0.28.0
pdf2image==1.16.3
Pillow==10.0.0
pytesseract==0.3.10
Requests==2.31.0
